app: update structure folder
