+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Added
 - ([#9](https://github.com/arrecadar/api/pull/9)) Dockerfile and docker-compose to development env.
 - Test runner, maybe [ava](https://github.com/avajs) with eslint-plugin.
-- Folder structure for `schema` with root queries and resolvers.
+- ([#11](https://github.com/arrecadar/api/pull/11)) Folder structure for `schema` with root queries and resolvers.
 - [Add build scripts before publish in production](https://github.com/babel/example-node-server)
 - [Add a healthCheck, we may need a workaround while this don't work properly](https://www.apollographql.com/docs/apollo-server/whats-new.html#Health-checks)
 - ([#10](https://github.com/arrecadar/api/pull/10)) Mongoose as ODM to handle mongodb
