+++ bootstrap/index.js
@@ -1,0 +1,2 @@
+export { default as App } from './app'
+export * from './database'
