+++ docker-compose.yml
@@ -7,6 +7,7 @@ services:
       target: development
     volumes:
       - ./bootstrap:/api/bootstrap
+      - ./schema/:/api/schema
       - ./server.js:/api/server.js
       - ./.babelrc:/api/.babelrc
